pyproject.toml
[project]
name = "shorts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.20.0",
    "praw>=7.8.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
